requirements.txt
# Requirements for YOLO label tools
Pillow>=8.0.0
